move termToRegex into its own file
