+++ packages/react-tei/src/unitex/enrichDocumentWithUnitex.ts
@@ -1,4 +1,5 @@
 import { mapTargetTags } from "../helper/mapTargetTags";
+import { escapeRegexChars, termToRegex } from "../helper/termToRegex";
 import type { DocumentJson } from "../parser/document";
 import {
 	type HighlightTag,
@@ -7,21 +8,6 @@ import {
 	type TextTag,
 } from "./highlightTermsInTextTag";
 import type { TermStatistic } from "./parseUnitexEnrichment";
-
-const escapeRegexChars = (str: string): string => {
-	return str
-		.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")
-		.replaceAll("-", "[-–]")
-		.replaceAll(/\s/g, "\\s");
-};
-
-export const termToRegex = (term: string): RegExp => {
-	const escapedTerm = escapeRegexChars(term);
-	return new RegExp(
-		`(?<![\\p{L}\\p{N}])${escapedTerm}(?![\\p{L}\\p{N}])`,
-		"gu",
-	);
-};
 
 type NormalizedTerm = {
 	term: string;
